feat: toHaveAccessibleName matcher
