--- src/helpers/matchers/accessibilityName.ts
@@ -1,18 +1,0 @@
-import { ReactTestInstance } from 'react-test-renderer';
-import { getAccessibleName } from '../../matchers/to-have-accessible-name';
-import { TextMatchOptions, matches, TextMatch } from '../../matches';
-
-export function matchAccessibleName(
-  node: ReactTestInstance,
-  expectedName?: TextMatch,
-  normalizer?: TextMatchOptions['normalizer'],
-  exact?: TextMatchOptions['exact']
-): boolean {
-  const accessibleName = getAccessibleName(node);
-
-  if (expectedName) {
-    return matches(expectedName, accessibleName, normalizer, exact);
-  }
-
-  return !!accessibleName;
-}
